test: Verify PR-562 resolves exceptiosn
This branch has the PR-562 merged.  It can be used to confirm that this issue no longer exists while the root cause of the bug still existed in release/0.10.
I am adding this branch in the event we want to verify what is stated in the Jira MTT-529 bug that PR-562 handles conditions where there is no NetworkManager around but the Network Update Loop is still invoking its NetworkUpdate method and that under these conditions the RpcQueueContainer and RpcQueueProcessor will not throw exceptions.